% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mazeGame.R
\name{maze}
\alias{maze}
\title{Maze Game}
\usage{
maze(size, cex_set = 2.5)
}
\arguments{
\item{size}{total maze size}

\item{cex_set}{each cell size}
}
\description{
Maze game, you can set the size by yourself.
}
\examples{
maze(25,2.5)

}
